# archive/litgo6_remaining_files/openapi.yaml
openapi: 3.0.3
info:
  title: LITGO5 API
  description: |
    API do Sistema de Matching Jurídico Inteligente LITGO5
    
    ## Visão Geral
    O LITGO5 é uma plataforma que conecta clientes a advogados especializados através de algoritmos de IA e análise semântica.
    
    ## Autenticação
    Todos os endpoints (exceto health check) requerem autenticação JWT via header `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Triagem: 30 requests/minuto
    - Conversação: 60 requests/minuto
    - Status: 120 requests/minuto
    - Outros endpoints: 100 requests/minuto
    
    ## Versioning
    Esta API segue versionamento semântico. Versão atual: v1
    
  version: 1.0.0
  contact:
    name: LITGO5 Team
    email: dev@litgo5.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Desenvolvimento Local
  - url: https://api.litgo5.com
    description: Produção
...
paths:
...
  /api/triage:
    post:
      tags:
        - Triagem
      summary: Triagem Inteligente Assíncrona
      description: Inicia processo de triagem assíncrona de caso com IA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - texto_cliente
                - coords
              properties:
                texto_cliente:
                  type: string
                  example: "Fui demitido sem justa causa e não recebi as verbas rescisórias"
                coords:
                  type: array
                  items:
                    type: number
                  example: [-23.5505, -46.6333]
      responses:
        '202':
          description: Triagem iniciada
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    example: "abc123-def456"
                  status:
                    type: string
                    example: "accepted"
                  message:
                    type: string
                    example: "A triagem do seu caso foi iniciada..."
  /api/triage/status/{task_id}:
    get:
      tags:
        - Triagem
      summary: Status da Triagem
      description: Consulta o status de uma triagem em andamento
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          example: "abc123-def456"
      responses:
        '200':
          description: Status da triagem
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, processing, completed, failed]
                  result:
                    type: object
                    properties:
                      case_id:
                        type: string
                      area:
                        type: string
                      subarea:
                        type: string
                      urgency_h:
                        type: integer
                      embedding:
                        type: array
                        items:
                          type: number
...
  /api/cases/my-cases:
    get:
      tags:
        - Cases
      summary: Meus Casos
      description: Lista todos os casos do usuário atual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de casos
          content:
            application/json:
              schema:
                type: object
                properties:
                  cases:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaseSummary'
  /api/cases/{case_id}:
    get:
      tags:
        - Cases
      summary: Detalhes do Caso
      description: Retorna detalhes completos de um caso específico
      security:
        - bearerAuth: []
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
          example: "case-789"
      responses:
        '200':
          description: Detalhes do caso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetails'
        '404':
          description: Caso não encontrado
        '403':
          description: Sem permissão para acessar este caso
  /api/cases/{case_id}/status:
    patch:
      tags:
        - Cases
      summary: Atualizar Status do Caso
      description: Atualiza o status de um caso
      security:
        - bearerAuth: []
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - new_status
              properties:
                new_status:
                  type: string
                  enum: [triagem, matching, offer_sent, in_progress, completed, cancelled]
                notes:
                  type: string
      responses:
        '200':
          description: Status atualizado com sucesso
        '400':
          description: Status inválido
        '403':
          description: Sem permissão para atualizar este caso
...
  /api/documents/case/{case_id}:
    get:
      tags:
        - Documents
      summary: Listar Documentos do Caso
      description: Lista todos os documentos de um caso
      security:
        - bearerAuth: []
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Lista de documentos
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/DocumentResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CaseSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        area:
          type: string
        subarea:
          type: string
        status:
          type: string
        lawyer_id:
          type: string
        lawyer_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        estimated_cost:
          type: number
        progress:
          type: integer
        next_deadline:
          type: object
          properties:
            date:
              type: string
              format: date-time
            description:
              type: string
        stats:
          type: object
          properties:
            messages_count:
              type: integer
            documents_count:
              type: integer
            events_count:
              type: integer
    CaseDetails:
      allOf:
        - $ref: '#/components/schemas/CaseSummary'
        - type: object
          properties:
            description:
              type: string
            client_id:
              type: string
            client_name:
              type: string
            contract_id:
              type: string
            final_fee:
              type: number
            service_scope:
              type: object
              properties:
                description:
                  type: string
                tasks:
                  type: array
                  items:
                    type: string
            timeline:
              type: object
              properties:
                created:
                  type: string
                  format: date-time
                lawyer_assigned:
                  type: string
                  format: date-time
                contract_signed:
                  type: string
                  format: date-time
                in_progress:
                  type: string
                  format: date-time
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
        case_id:
          type: string
        uploaded_by:
          type: string
        file_name:
          type: string
        file_size:
          type: integer
        file_type:
          type: string
        file_url:
          type: string
        storage_path:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...